Update: Move tests closer to implementation
